rag/requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.6.1
faiss-cpu>=1.8.0
numpy>=1.26.0
